Category verification for exclusion

+++ app/Http/Controllers/Admin/ProductController.php
@@ -111,6 +111,13 @@ class ProductController extends Controller
      */
     public function destroy($id)
     {
+        $products = count($this->repository->productsByCategoryId($id));
+        if(count($products) > 0) {
+            return redirect()
+                   ->route('categories.index')
+                   ->with('message', 'Não pode deletar porque existem produtos vinculados a essa categoria!');
+        }
+
         $this->repository->delete($id);
 
         return redirect()
